pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["PashkovichKaterina <pashkovichkate@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.1"
psycopg2-binary = "^2.9.3"
djangorestframework = "^3.13.1"
drf-spectacular = "^0.21.2"
django-cors-headers = "^3.11.0"
djangorestframework-simplejwt = "^5.0.0"
boto3 = "^1.21.0"
django-storages = "^1.12.3"
drf-nested-routers = "^0.93.4"
factory-boy = "^3.2.1"
moto = {extras = ["s3"], version = "^3.0.3"}
django-anymail = {extras = ["sendinblue"], version = "^8.5"}
celery = "^5.2.5"
redis = "^4.2.1"
gunicorn = "^20.1.0"
watchtower = "^3.0.0"
launchdarkly-server-sdk = "^7.4.0"
SpeechRecognition = "^3.8.1"
pydub = "^0.25.1"
django-lifecycle = "^0.9.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
